refactor(nests): split nests/room flat package into named subpackages
The nests/room/ package had grown to 23+ files in one flat directory,
making it hard to find related composables and lifecycle helpers. Each
file is now grouped by responsibility:

  activity/     Android Activity entry + AccountViewModel bridge
  lobby/        Pre-join lobby card (NestJoinCard, JoinNestButton)
  lifecycle/    Side-effect composables wiring VM <-> system / cache
  screen/       Top-level layouts (FullScreen, PipScreen, ActionBar)
  stage/        Speaker / audience grids and overlays
  participants/ Per-participant host actions + role mutations
  reactions/    Reaction picker
  chat/         Chat panel and composer (also absorbs the old send/)
  edit/         Edit-room sheet + ViewModel
  theme/        Themed scope + URL font loader

Also drops the unused StagePeopleRow composable from NestCommon and
inlines its remaining one-liner connectingLabel helper into
NestActionBar (the only caller), removing NestCommon entirely.

External imports in 7 sibling files were updated to point at the new
subpackage paths.
